# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Customer API Documentation",
        "description": "Documentation de l'API Customer",
        "contact": {
            "email": "support@example.com"
        },
        "version": "2.0.0"
    },
    "paths": {
        "/customer-addresses": {
            "get": {
                "tags": [
                    "CustomerAddresses"
                ],
                "summary": "Liste toutes les adresses clients",
                "operationId": "df466a4bec3008e7972aa803f65ca464",
                "responses": {
                    "200": {
                        "description": "Liste des adresses clients récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/CustomerAddress"
                                            }
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Liste des adresses clients récupérée avec succès."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "CustomerAddresses"
                ],
                "summary": "Crée une nouvelle adresse client",
                "operationId": "9f09f924361a46d9cb95743b472c4fcd",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreCustomerAddressRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Adresse client créée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/CustomerAddress"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Adresse client créée avec succès."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                }
            }
        },
        "/customer-addresses/{id}": {
            "get": {
                "tags": [
                    "CustomerAddresses"
                ],
                "summary": "Affiche les détails d'une adresse client spécifique",
                "operationId": "dfe50df677b6603e3de66df646b11f50",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'adresse client",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Adresse client récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/CustomerAddress"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Adresse client récupérée avec succès."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Adresse client non trouvée"
                    }
                }
            },
            "put": {
                "tags": [
                    "CustomerAddresses"
                ],
                "summary": "Met à jour une adresse client existante",
                "operationId": "33d0ef10fdd3a59e0fba461bea3035ca",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'adresse client à modifier",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateCustomerAddressRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Adresse client mise à jour avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/CustomerAddress"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Adresse client mise à jour avec succès."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Adresse client non trouvée"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                }
            },
            "delete": {
                "tags": [
                    "CustomerAddresses"
                ],
                "summary": "Supprime une adresse client",
                "operationId": "d6cfbfa8dbda8876b396a8231442174f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'adresse client à supprimer",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Adresse client supprimée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Adresse client supprimée avec succès."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Adresse client non trouvée"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "StoreCustomerAddressRequest": {
                "required": [
                    "first_name",
                    "last_name",
                    "address",
                    "postal_code",
                    "city",
                    "id_country",
                    "phone"
                ],
                "properties": {
                    "first_name": {
                        "description": "Prénom",
                        "type": "string",
                        "maxLength": 64,
                        "example": "Jean"
                    },
                    "last_name": {
                        "description": "Nom",
                        "type": "string",
                        "maxLength": 64,
                        "example": "Dupont"
                    },
                    "address": {
                        "description": "Adresse",
                        "type": "string",
                        "maxLength": 255,
                        "example": "123 Rue de la République"
                    },
                    "complement_address": {
                        "description": "Complément d'adresse",
                        "type": "string",
                        "maxLength": 200,
                        "example": "Bâtiment A, étage 2",
                        "nullable": true
                    },
                    "postal_code": {
                        "description": "Code postal",
                        "type": "string",
                        "maxLength": 10,
                        "example": "75001"
                    },
                    "city": {
                        "description": "Ville",
                        "type": "string",
                        "maxLength": 64,
                        "example": "Paris"
                    },
                    "id_country": {
                        "description": "ID du pays",
                        "type": "integer",
                        "example": 1
                    },
                    "phone": {
                        "description": "Téléphone",
                        "type": "string",
                        "maxLength": 20,
                        "example": "+33123456789"
                    },
                    "fax": {
                        "description": "Fax",
                        "type": "string",
                        "maxLength": 20,
                        "example": "+33123456780",
                        "nullable": true
                    },
                    "longitude": {
                        "description": "Longitude",
                        "type": "number",
                        "format": "float",
                        "example": 2.3522219,
                        "nullable": true
                    },
                    "latitude": {
                        "description": "Latitude",
                        "type": "number",
                        "format": "float",
                        "example": 48.856614,
                        "nullable": true
                    },
                    "place_id": {
                        "description": "ID de lieu Google Maps",
                        "type": "string",
                        "maxLength": 150,
                        "example": "ChIJD7fiBh9u5kcRYJSMaMOCCwQ",
                        "nullable": true
                    },
                    "address_name": {
                        "description": "Nom personnalisé de l'adresse",
                        "type": "string",
                        "maxLength": 64,
                        "example": "Domicile",
                        "nullable": true
                    },
                    "has_difficult_access": {
                        "description": "Accès difficile",
                        "type": "boolean",
                        "example": false
                    }
                },
                "type": "object"
            },
            "UpdateCustomerAddressRequest": {
                "properties": {
                    "first_name": {
                        "description": "Prénom",
                        "type": "string",
                        "maxLength": 64,
                        "example": "Jean"
                    },
                    "last_name": {
                        "description": "Nom",
                        "type": "string",
                        "maxLength": 64,
                        "example": "Dupont"
                    },
                    "address": {
                        "description": "Adresse",
                        "type": "string",
                        "maxLength": 255,
                        "example": "123 Rue de la République"
                    },
                    "complement_address": {
                        "description": "Complément d'adresse",
                        "type": "string",
                        "maxLength": 200,
                        "example": "Bâtiment A, étage 2",
                        "nullable": true
                    },
                    "postal_code": {
                        "description": "Code postal",
                        "type": "string",
                        "maxLength": 10,
                        "example": "75001"
                    },
                    "city": {
                        "description": "Ville",
                        "type": "string",
                        "maxLength": 64,
                        "example": "Paris"
                    },
                    "id_country": {
                        "description": "ID du pays",
                        "type": "integer",
                        "example": 1
                    },
                    "phone": {
                        "description": "Téléphone",
                        "type": "string",
                        "maxLength": 20,
                        "example": "+33123456789"
                    },
                    "fax": {
                        "description": "Fax",
                        "type": "string",
                        "maxLength": 20,
                        "example": "+33123456780",
                        "nullable": true
                    },
                    "longitude": {
                        "description": "Longitude",
                        "type": "number",
                        "format": "float",
                        "example": 2.3522219,
                        "nullable": true
                    },
                    "latitude": {
                        "description": "Latitude",
                        "type": "number",
                        "format": "float",
                        "example": 48.856614,
                        "nullable": true
                    },
                    "place_id": {
                        "description": "ID de lieu Google Maps",
                        "type": "string",
                        "maxLength": 150,
                        "example": "ChIJD7fiBh9u5kcRYJSMaMOCCwQ",
                        "nullable": true
                    },
                    "address_name": {
                        "description": "Nom personnalisé de l'adresse",
                        "type": "string",
                        "maxLength": 64,
                        "example": "Domicile",
                        "nullable": true
                    },
                    "has_difficult_access": {
                        "description": "Accès difficile",
                        "type": "boolean",
                        "example": false
                    }
                },
                "type": "object"
            },
...
            "CustomerAddress": {
                "title": "CustomerAddress",
                "description": "Modèle d'adresse client",
                "properties": {
                    "id_customer_address": {
                        "description": "ID unique de l'adresse client",
                        "type": "integer",
                        "example": 1
                    },
                    "first_name": {
                        "description": "Prénom",
                        "type": "string",
                        "maxLength": 64,
                        "example": "Jean"
                    },
                    "last_name": {
                        "description": "Nom",
                        "type": "string",
                        "maxLength": 64,
                        "example": "Dupont"
                    },
                    "address": {
                        "description": "Adresse",
                        "type": "string",
                        "maxLength": 255,
                        "example": "123 Rue de la République"
                    },
                    "complement_address": {
                        "description": "Complément d'adresse",
                        "type": "string",
                        "maxLength": 200,
                        "example": "Bâtiment A, étage 2",
                        "nullable": true
                    },
                    "postal_code": {
                        "description": "Code postal",
                        "type": "string",
                        "maxLength": 10,
                        "example": "75001"
                    },
                    "city": {
                        "description": "Ville",
                        "type": "string",
                        "maxLength": 64,
                        "example": "Paris"
                    },
                    "id_country": {
                        "description": "ID du pays",
                        "type": "integer",
                        "example": 1
                    },
                    "phone": {
                        "description": "Téléphone",
                        "type": "string",
                        "maxLength": 20,
                        "example": "+33123456789"
                    },
                    "fax": {
                        "description": "Fax",
                        "type": "string",
                        "maxLength": 20,
                        "example": "+33123456780",
                        "nullable": true
                    },
                    "longitude": {
                        "description": "Longitude",
                        "type": "number",
                        "format": "float",
                        "example": 2.3522219,
                        "nullable": true
                    },
                    "latitude": {
                        "description": "Latitude",
                        "type": "number",
                        "format": "float",
                        "example": 48.856614,
                        "nullable": true
                    },
                    "place_id": {
                        "description": "ID de lieu Google Maps",
                        "type": "string",
                        "maxLength": 150,
                        "example": "ChIJD7fiBh9u5kcRYJSMaMOCCwQ",
                        "nullable": true
                    },
                    "address_name": {
                        "description": "Nom personnalisé de l'adresse",
                        "type": "string",
                        "maxLength": 64,
                        "example": "Domicile",
                        "nullable": true
                    },
                    "has_difficult_access": {
                        "description": "Accès difficile",
                        "type": "boolean",
                        "example": false
                    },
                    "created_at": {
                        "description": "Date de création",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Date de dernière modification",
                        "type": "string",
                        "format": "date-time"
                    },
                    "country": {
                        "$ref": "#/components/schemas/CustomerCountry"
                    }
                },
                "type": "object"
            },
...
            "CustomerCountry": {
                "title": "CustomerCountry",
                "description": "Modèle de pays client",
                "properties": {
                    "id_customer_country": {
                        "description": "ID unique du pays client",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "Nom du pays en français",
                        "type": "string",
                        "maxLength": 200,
                        "example": "France"
                    },
                    "name_en": {
                        "description": "Nom du pays en anglais",
                        "type": "string",
                        "maxLength": 70,
                        "example": "France"
                    },
                    "name_de": {
                        "description": "Nom du pays en allemand",
                        "type": "string",
                        "maxLength": 70,
                        "example": "Frankreich"
                    },
                    "isocode": {
                        "description": "Code ISO du pays",
                        "type": "string",
                        "maxLength": 5,
                        "example": "FR"
                    },
                    "ccn3": {
                        "description": "Code numérique du pays",
                        "type": "integer",
                        "example": 250,
                        "nullable": true
                    },
                    "phone_code": {
                        "description": "Indicatif téléphonique",
                        "type": "string",
                        "maxLength": 10,
                        "example": "+33",
                        "nullable": true
                    },
                    "is_intracom_vat": {
                        "description": "Pays intracommunautaire pour TVA (0=non, 1=oui)",
                        "type": "integer",
                        "example": 1
                    },
                    "is_ue_export": {
                        "description": "Pays UE pour export (0=non, 1=oui)",
                        "type": "integer",
                        "example": 1
                    },
                    "created_at": {
                        "description": "Date de création",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Date de dernière modification",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}